Clarify Reference to Fingerprint Verification Methods
There was originally a section here about using questions to verify someone in lieu of fingerprint verification, but the way the section was structured (including references to an MITM attack still being possible if your conversation partner knows things only she would know) seemed to undermine the legitimacy of these questions.

Now, the section maintains that fingerprint verification's the best way to verify contacts (especially on Adium!) but on-the-fly methods (like Micah tweeting Snowden's fingerprint) are still decent. (My thinking is that it was rock solid in that case but won't always be -- not all of us have 11k followers on Twitter and a key signed by Cory Doctorow...)