Correction: signatures versus checksums
Original text equated checksums and signatures, which are two different things and processes (it referred to checksums but then gave the process for verifying a signature, and used them interchangeably). A checksum is a cryptographic hash that's unique to the exact contents of the file. A signature is an assertion about the checksum of a file by a public key (i.e. the developers). The edit now mentions both.