Separates title and subtitle.
The title of the document should be formatted as h1.
Some may disagree about formatting the subtitle as h2,
since it's not a structural distinction, but rather
a formatting aid for readability. The rationale is that
the main title ("Encryption Works") should be as prominent
as possible, since that is the title most folks will use
when discussing the piece or referencing it.