Remove ambiguous sentence from key signing section
As pointed out in #195, there is a sentence which seems totally out of place
here and it doesn't seem like it belongs in any of the nearby paragraphs as
well. It might be an artifact from a deleted section or paragraph. Anyway, this
commit removes it. Note to close #195 when this is merged.

Signed-off-by: Noah Vesely <fowlslegs@riseup.net>