Confidentiality, anonymity, and getting more of the latter
First, this adds two paragraphs on confidentiality vs. anonymity to the "Threat
Modeling" section.

Then, in the Signal and OTR sections it adds information on how these
technologies do not provide anonymity. In the OTR section, there was already
some information to this regard. I've moved it into it's own subsection and
added to it some.

In this new "OTR, Metadata, and Anonymity" subsection, I've also added some
information on how to use OTR over Tor on Android devices by using ChatSecure
together w/ Orbot.

Most of this was discussed in #214.

Signed-off-by: Noah Vesely <fowlslegs@riseup.net>